+++ crates/cli/tests/reference/echo.js
@@ -725,7 +725,7 @@ export function echo_vec_struct(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_u8(a) {
@@ -734,7 +734,7 @@ export function echo_option_u8(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_i8(a) {
@@ -743,7 +743,7 @@ export function echo_option_i8(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_u16(a) {
@@ -752,7 +752,7 @@ export function echo_option_u16(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_i16(a) {
@@ -761,7 +761,7 @@ export function echo_option_i16(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_u32(a) {
@@ -770,7 +770,7 @@ export function echo_option_u32(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_i32(a) {
@@ -779,7 +779,7 @@ export function echo_option_i32(a) {
 }
 
 /**
- * @param {bigint | undefined} [a]
+ * @param {bigint | null} [a]
  * @returns {bigint | undefined}
  */
 export function echo_option_u64(a) {
@@ -788,7 +788,7 @@ export function echo_option_u64(a) {
 }
 
 /**
- * @param {bigint | undefined} [a]
+ * @param {bigint | null} [a]
  * @returns {bigint | undefined}
  */
 export function echo_option_i64(a) {
@@ -797,7 +797,7 @@ export function echo_option_i64(a) {
 }
 
 /**
- * @param {bigint | undefined} [a]
+ * @param {bigint | null} [a]
  * @returns {bigint | undefined}
  */
 export function echo_option_u128(a) {
@@ -806,7 +806,7 @@ export function echo_option_u128(a) {
 }
 
 /**
- * @param {bigint | undefined} [a]
+ * @param {bigint | null} [a]
  * @returns {bigint | undefined}
  */
 export function echo_option_i128(a) {
@@ -815,7 +815,7 @@ export function echo_option_i128(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_usize(a) {
@@ -824,7 +824,7 @@ export function echo_option_usize(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_isize(a) {
@@ -833,7 +833,7 @@ export function echo_option_isize(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_f32(a) {
@@ -842,7 +842,7 @@ export function echo_option_f32(a) {
 }
 
 /**
- * @param {number | undefined} [a]
+ * @param {number | null} [a]
  * @returns {number | undefined}
  */
 export function echo_option_f64(a) {
@@ -851,7 +851,7 @@ export function echo_option_f64(a) {
 }
 
 /**
- * @param {boolean | undefined} [a]
+ * @param {boolean | null} [a]
  * @returns {boolean | undefined}
  */
 export function echo_option_bool(a) {
@@ -860,7 +860,7 @@ export function echo_option_bool(a) {
 }
 
 /**
- * @param {string | undefined} [a]
+ * @param {string | null} [a]
  * @returns {string | undefined}
  */
 export function echo_option_char(a) {
@@ -871,7 +871,7 @@ export function echo_option_char(a) {
 }
 
 /**
- * @param {string | undefined} [a]
+ * @param {string | null} [a]
  * @returns {string | undefined}
  */
 export function echo_option_string(a) {
@@ -887,7 +887,7 @@ export function echo_option_string(a) {
 }
 
 /**
- * @param {Uint8Array | undefined} [a]
+ * @param {Uint8Array | null} [a]
  * @returns {Uint8Array | undefined}
  */
 export function echo_option_vec_u8(a) {
@@ -903,7 +903,7 @@ export function echo_option_vec_u8(a) {
 }
 
 /**
- * @param {Int8Array | undefined} [a]
+ * @param {Int8Array | null} [a]
  * @returns {Int8Array | undefined}
  */
 export function echo_option_vec_i8(a) {
@@ -919,7 +919,7 @@ export function echo_option_vec_i8(a) {
 }
 
 /**
- * @param {Uint16Array | undefined} [a]
+ * @param {Uint16Array | null} [a]
  * @returns {Uint16Array | undefined}
  */
 export function echo_option_vec_u16(a) {
@@ -935,7 +935,7 @@ export function echo_option_vec_u16(a) {
 }
 
 /**
- * @param {Int16Array | undefined} [a]
+ * @param {Int16Array | null} [a]
  * @returns {Int16Array | undefined}
  */
 export function echo_option_vec_i16(a) {
@@ -951,7 +951,7 @@ export function echo_option_vec_i16(a) {
 }
 
 /**
- * @param {Uint32Array | undefined} [a]
+ * @param {Uint32Array | null} [a]
  * @returns {Uint32Array | undefined}
  */
 export function echo_option_vec_u32(a) {
@@ -967,7 +967,7 @@ export function echo_option_vec_u32(a) {
 }
 
 /**
- * @param {Int32Array | undefined} [a]
+ * @param {Int32Array | null} [a]
  * @returns {Int32Array | undefined}
  */
 export function echo_option_vec_i32(a) {
@@ -983,7 +983,7 @@ export function echo_option_vec_i32(a) {
 }
 
 /**
- * @param {BigUint64Array | undefined} [a]
+ * @param {BigUint64Array | null} [a]
  * @returns {BigUint64Array | undefined}
  */
 export function echo_option_vec_u64(a) {
@@ -999,7 +999,7 @@ export function echo_option_vec_u64(a) {
 }
 
 /**
- * @param {BigInt64Array | undefined} [a]
+ * @param {BigInt64Array | null} [a]
  * @returns {BigInt64Array | undefined}
  */
 export function echo_option_vec_i64(a) {
@@ -1015,7 +1015,7 @@ export function echo_option_vec_i64(a) {
 }
 
 /**
- * @param {Uint8Array | undefined} [a]
+ * @param {Uint8Array | null} [a]
  * @returns {Uint8Array | undefined}
  */
 export function echo_option_vec_uninit_u8(a) {
@@ -1031,7 +1031,7 @@ export function echo_option_vec_uninit_u8(a) {
 }
 
 /**
- * @param {Int8Array | undefined} [a]
+ * @param {Int8Array | null} [a]
  * @returns {Int8Array | undefined}
  */
 export function echo_option_vec_uninit_i8(a) {
@@ -1047,7 +1047,7 @@ export function echo_option_vec_uninit_i8(a) {
 }
 
 /**
- * @param {Uint16Array | undefined} [a]
+ * @param {Uint16Array | null} [a]
  * @returns {Uint16Array | undefined}
  */
 export function echo_option_vec_uninit_u16(a) {
@@ -1063,7 +1063,7 @@ export function echo_option_vec_uninit_u16(a) {
 }
 
 /**
- * @param {Int16Array | undefined} [a]
+ * @param {Int16Array | null} [a]
  * @returns {Int16Array | undefined}
  */
 export function echo_option_vec_uninit_i16(a) {
@@ -1079,7 +1079,7 @@ export function echo_option_vec_uninit_i16(a) {
 }
 
 /**
- * @param {Uint32Array | undefined} [a]
+ * @param {Uint32Array | null} [a]
  * @returns {Uint32Array | undefined}
  */
 export function echo_option_vec_uninit_u32(a) {
@@ -1095,7 +1095,7 @@ export function echo_option_vec_uninit_u32(a) {
 }
 
 /**
- * @param {Int32Array | undefined} [a]
+ * @param {Int32Array | null} [a]
  * @returns {Int32Array | undefined}
  */
 export function echo_option_vec_uninit_i32(a) {
@@ -1111,7 +1111,7 @@ export function echo_option_vec_uninit_i32(a) {
 }
 
 /**
- * @param {BigUint64Array | undefined} [a]
+ * @param {BigUint64Array | null} [a]
  * @returns {BigUint64Array | undefined}
  */
 export function echo_option_vec_uninit_u64(a) {
@@ -1127,7 +1127,7 @@ export function echo_option_vec_uninit_u64(a) {
 }
 
 /**
- * @param {BigInt64Array | undefined} [a]
+ * @param {BigInt64Array | null} [a]
  * @returns {BigInt64Array | undefined}
  */
 export function echo_option_vec_uninit_i64(a) {
@@ -1143,7 +1143,7 @@ export function echo_option_vec_uninit_i64(a) {
 }
 
 /**
- * @param {(string)[] | undefined} [a]
+ * @param {(string)[] | null} [a]
  * @returns {(string)[] | undefined}
  */
 export function echo_option_vec_string(a) {
@@ -1159,7 +1159,7 @@ export function echo_option_vec_string(a) {
 }
 
 /**
- * @param {Foo | undefined} [a]
+ * @param {Foo | null} [a]
  * @returns {Foo | undefined}
  */
 export function echo_option_struct(a) {
@@ -1173,7 +1173,7 @@ export function echo_option_struct(a) {
 }
 
 /**
- * @param {(Foo)[] | undefined} [a]
+ * @param {(Foo)[] | null} [a]
  * @returns {(Foo)[] | undefined}
  */
 export function echo_option_vec_struct(a) {