+++ crates/cli/tests/reference/getter-setter.js
@@ -124,7 +124,7 @@ export class Foo {
         return ret === 0x100000001 ? undefined : ret;
     }
     /**
-     * @param {number | undefined} [arg0]
+     * @param {number | null} [arg0]
      */
     set y(arg0) {
         wasm.__wbg_set_foo_y(this.__wbg_ptr, isLikeNone(arg0) ? 0x100000001 : (arg0) >>> 0);
@@ -137,7 +137,7 @@ export class Foo {
         return ret === 0x100000001 ? undefined : ret;
     }
     /**
-     * @param {number | undefined} [z]
+     * @param {number | null} [z]
      */
     set z(z) {
         wasm.foo_set_z(this.__wbg_ptr, isLikeNone(z) ? 0x100000001 : (z) >>> 0);
@@ -150,7 +150,7 @@ export class Foo {
         return ret === 0x100000001 ? undefined : ret;
     }
     /**
-     * @param {number | undefined} [value]
+     * @param {number | null} [value]
      */
     set lone_setter(value) {
         wasm.foo_set_lone_setter(this.__wbg_ptr, isLikeNone(value) ? 0x100000001 : (value) >>> 0);
@@ -167,7 +167,7 @@ export class Foo {
      * But you must write strings.
      *
      * Yes, this is totally fine in JS.
-     * @param {string | undefined} [value]
+     * @param {string | null} [value]
      */
     set weird(value) {
         const ptr = isLikeNone(value) ? 0 : passStringToWasm0(value, wasm.__wbindgen_malloc, wasm.__wbindgen_realloc);
@@ -184,7 +184,7 @@ export class Foo {
         return ret === 0xFFFFFF ? undefined : ret !== 0;
     }
     /**
-     * @param {boolean | undefined} [value]
+     * @param {boolean | null} [value]
      */
     static set x(value) {
         wasm.foo_set_x_static(isLikeNone(value) ? 0xFFFFFF : value ? 1 : 0);